requirements.txt
Package                Version
---------------------- -----------
Django                 4.1.2
django-filter          22.1
djangorestframework    3.14.0
Faker                  15.1.1
Jinja2                 3.1.2
Markdown               3.4.1
mysql-connector        2.2.9
mysql-connector-python 8.0.31
mysqlclient            2.1.1
numpy                  1.23.4
openpyxl               3.0.10
pandas                 1.5.1
pip                    22.3.1
pyfarmhash             0.3.2
SQLAlchemy             1.4.42